feat: added version command
