+++ src/sophrosyne/core/models.py
@@ -16,7 +16,7 @@ validated against these models and schemas.
 
 from datetime import datetime
 from enum import Enum
-from typing import Annotated, Literal, Union
+from typing import Union
 
 from pydantic import BaseModel, EmailStr
 from sqlmodel import (
